Changes to Course Syllabus
Course Structure Change: The change in this section is from a seminar
to a lecture format. I think there is greater value in learning from
the instructor's vast knowledge of the subject rather than the Socratic
teeth pulling of a
room full of earnest but uninformed students.

Deliverables: The change in this section is from a paper to a software
evaluation assignment. Plus removing some of the weight from the memos.
This is due to a change in the memo requirements listed below.

Participation: The change in this section is a reflection of the shift
from a seminar format to a lecture format. Participation will still be
of value, hopefully informing the direction of future lectures.

Memos: There are a few changes in this section. The first is a
reduction of the amount of memos from nine to five and allowing the
students to select which five classes they will submit memos on.
Perhaps this is done at the first or second class and then groups of
2-3 students are formed and they lead the initial part of the
discussion that day. This layout allows students to select times that
work better for them and so that they do not feel micromanaged to do
their readings.

The major assignment changes were a shift from a paper to a business
case software proposal. This would still allow students to look at the
theory and issues surrounding the software while giving practical hands
on experience.